backend/commons/security/security-authentication-noop/src/main/java/com/kraken/security/authentication/noop/NoopExchangeFilter.java
package com.kraken.security.authentication.noop;

import com.kraken.security.authentication.api.ExchangeFilter;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Component;
import org.springframework.web.reactive.function.client.ClientRequest;
import org.springframework.web.reactive.function.client.ClientResponse;
import org.springframework.web.reactive.function.client.ExchangeFunction;
import reactor.core.publisher.Mono;

@Slf4j
@Component
final class NoopExchangeFilter implements ExchangeFilter {

  @Override
  public Mono<ClientResponse> filter(final ClientRequest request, final ExchangeFunction next) {
    return next.exchange(request);
  }

}
